Moves CIDR handling functions to `internal/types` instead of `internal/verify`
