+++ internal/service/ec2/diff.go
@@ -2,11 +2,11 @@ package ec2
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-provider-aws/internal/verify"
+	"github.com/hashicorp/terraform-provider-aws/internal/types"
 )
 
 // suppressEqualCIDRBlockDiffs provides custom difference suppression for CIDR blocks
 // that have different string values but represent the same CIDR.
 func suppressEqualCIDRBlockDiffs(k, old, new string, d *schema.ResourceData) bool {
-	return verify.CIDRBlocksEqual(old, new)
+	return types.CIDRBlocksEqual(old, new)
 }
